Changed data types in CRUD class and interface

+++ GIBZ_M151_ToDoList_2021/L_DataAccess/Functions/Interfaces/ICRUD.cs
@@ -9,9 +9,9 @@ namespace L_DataAccess.Functions.Interfaces
     public interface ICRUD
     {
         Task<T> Create<T>(T objectForDb) where T : class;
-        Task<T> Read<T>(Int64 entityId) where T : class;
+        Task<T> Read<T>(int entityId) where T : class;
         Task<List<T>> ReadAll<T>() where T : class;
-        Task<T> Update<T>(T objectToUpdate, Int64 entityId) where T : class;
-        Task<bool> Delete<T>(Int64 entityId) where T : class;
+        Task<T> Update<T>(T objectToUpdate, int entityId) where T : class;
+        Task<bool> Delete<T>(int entityId) where T : class;
     }
 }
